Patch for gcc7

+++ 3rd-party/install/Elemental/link.sh
@@ -7,11 +7,13 @@ set_prefix
 
 BUILD_TYPES="Release Debug"
 for build_type in $BUILD_TYPES; do
-  $SUDO rm -f $PREFIX_ROKKO/elemental
-  $SUDO ln -s $PREFIX_ROKKO/elemental-$ELEMENTAL_VERSION-$ELEMENTAL_RK_REVISION $PREFIX_ROKKO/elemental
-  $SUDO rm -f $PREFIX_ROKKO/rokkoenv-$build_type.d/elementalvars.sh
-  $SUDO ln -s $PREFIX_ROKKO/elemental/$build_type/elementalvars.sh $PREFIX_ROKKO/rokkoenv-$build_type.d/elementalvars.sh
+  if [ -f $PREFIX_ROKKO/elemental/elemental-$ELEMENTAL_VERSION-$ELEMENTAL_RK_REVISION/$build_type/elementalvars.sh ]; then
+    $SUDO rm -f $PREFIX_ROKKO/rokkoenv-$build_type.d/elementalvars.sh
+    $SUDO ln -s $PREFIX_ROKKO/elemental/elemental-$ELEMENTAL_VERSION-$ELEMENTAL_RK_REVISION/$build_type/elementalvars.sh $PREFIX_ROKKO/rokkoenv-$build_type.d/elementalvars.sh
+  fi
 done
 
-$SUDO rm -f $PREFIX_ROKKO/rokkoenv.d/elementalvars.sh
-$SUDO ln -s $PREFIX_ROKKO/elemental/elementalvars.sh $PREFIX_ROKKO/rokkoenv.d/elementalvars.sh
+if [ -f $PREFIX_ROKKO/elemental/elemental-$ELEMENTAL_VERSION-$ELEMENTAL_RK_REVISION/elementalvars.sh ]; then
+  $SUDO rm -f $PREFIX_ROKKO/rokkoenv.d/elementalvars.sh
+  $SUDO ln -s $PREFIX_ROKKO/elemental/elemental-$ELEMENTAL_VERSION-$ELEMENTAL_RK_REVISION/elementalvars.sh $PREFIX_ROKKO/rokkoenv.d/elementalvars.sh
+fi
